Oye..
All the required parts are done BUT I’m having some functionality
issues when it comes to the “no thanks” button. the pop up only flashes
up as opposed to staying until clicked(Which should also close the
window)…even when its not hidden from the start the functionality of
closing doesn’t work. But the ‘X’ closes just fine. Oh yeah… the @media
stuff is in. I’m not sure what depth I have to go into past the main
containers and sections. I tested it in Firefox’s Responsive Design
Mode and it all stays the same(i hope that was the goal)